refactor: reusable formatting tools
